move onboarding functions test
