+++ onboarding-functions/src/test/java/it/pagopa/selfcare/onboarding/HttpResponseMessageMock.java
@@ -1,4 +1,4 @@
-package it.pagopa.selfcare;
+package it.pagopa.selfcare.onboarding;
 
 import com.microsoft.azure.functions.HttpResponseMessage;
 import com.microsoft.azure.functions.HttpStatus;
